fix(#1, #5),feat(#3): fixed edge bug, removing data bug, and added file extensions
