+++ main.py
@@ -1,5 +1,5 @@
-from gui import App
+from gui.app import App
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     app = App()
     app.mainloop()
